requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
google-api-python-client==2.150.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
SQLAlchemy==2.0.44